+++ src/main/kotlin/logic/models/Project.kt
@@ -5,4 +5,5 @@ import java.util.*
 data class Project(
     val id: UUID,
     val name: String,
+    val userIds: List<UUID>
 )
