--- src/main/kotlin/logic/usecases/GetStateByIdUseCase.kt
@@ -1,13 +1,0 @@
-package logic.usecases
-
-import logic.models.TaskState
-import logic.repositoies.StateRepository
-import java.util.*
-
-class GetStateByIdUseCase(
-    private val stateRepository: StateRepository
-) {
-    operator fun invoke(id: UUID): TaskState{
-        return TaskState(UUID.randomUUID(), " ", UUID.randomUUID())
-    }
-}
