--- src/main/kotlin/logic/usecases/GetStatesByProjectIdUseCase.kt
@@ -1,14 +1,0 @@
-package logic.usecases
-
-import logic.models.TaskState
-import logic.repositoies.StateRepository
-import java.util.UUID
-
-class GetStatesByProjectIdUseCase(
-    private val stateRepository: StateRepository
-) {
-
-    operator fun invoke(projectId: UUID): List<TaskState>{
-        return emptyList()
-    }
-}
