+++ src/main/kotlin/logic/usecases/project/GetAllProjectsUseCase.kt
@@ -1,0 +1,13 @@
+package logic.usecases.project
+
+import logic.models.Project
+import logic.repositoies.ProjectsRepository
+
+class GetAllProjectsUseCase(
+    private val projectsRepository: ProjectsRepository
+) {
+
+    operator fun invoke(): List<Project> {
+        return projectsRepository.getAll()
+    }
+}
