+++ src/main/kotlin/logic/usecases/project/GetProjectByIdUseCase.kt
@@ -1,0 +1,14 @@
+package logic.usecases.project
+
+import logic.models.Project
+import logic.repositoies.ProjectsRepository
+import java.util.UUID
+
+class GetProjectByIdUseCase(
+    private val projectsRepository: ProjectsRepository
+) {
+
+    operator fun invoke(id: UUID): Project {
+        return projectsRepository.getById(id)
+    }
+}
