+++ src/main/kotlin/logic/usecases/task/AddTaskUseCase.kt
@@ -1,0 +1,12 @@
+package logic.usecases.task
+
+import logic.models.Task
+import logic.repositoies.TasksRepository
+
+class AddTaskUseCase(
+    private val tasksRepository: TasksRepository
+) {
+    operator fun invoke(task: Task): Boolean {
+        return tasksRepository.add(task)
+    }
+}
