+++ src/main/kotlin/logic/usecases/task/DeleteTaskUseCase.kt
@@ -1,0 +1,14 @@
+package logic.usecases.task
+
+import logic.repositoies.TasksRepository
+import java.util.*
+
+class DeleteTaskUseCase(
+    private val tasksRepository: TasksRepository
+) {
+
+    operator fun invoke(id: UUID): Boolean {
+        return tasksRepository.delete(id)
+    }
+
+}
