+++ src/main/kotlin/logic/usecases/task/GetTaskByIdUseCase.kt
@@ -1,0 +1,14 @@
+package logic.usecases.task
+
+import logic.models.Task
+import logic.repositoies.TasksRepository
+import java.util.*
+
+class GetTaskByIdUseCase(
+    private val tasksRepository: TasksRepository
+) {
+
+    operator fun invoke(id: UUID): Task {
+        return tasksRepository.getById(id)
+    }
+}
