+++ src/main/kotlin/logic/usecases/task/GetTaskByProjectIdUseCase.kt
@@ -1,0 +1,13 @@
+package logic.usecases.task
+
+import logic.models.Task
+import logic.repositoies.TasksRepository
+import java.util.*
+
+class GetTaskByProjectIdUseCase(
+    private val tasksRepository: TasksRepository
+) {
+    operator fun invoke(id: UUID): List<Task> {
+        return tasksRepository.getByProjectId(id)
+    }
+}
