--- src/main/kotlin/logic/usecases/tasksUseCases/DeleteTaskUseCase.kt
@@ -1,14 +1,0 @@
-package logic.usecases.tasksUseCases
-
-import logic.repositoies.TasksRepository
-import java.util.*
-
-class DeleteTaskUseCase(
-    private val tasksRepository: TasksRepository
-) {
-
-    operator fun invoke(id: UUID): Boolean {
-        return tasksRepository.delete(id)
-    }
-
-}
