--- src/main/kotlin/logic/usecases/tasksUseCases/EditTaskUseCase.kt
@@ -1,12 +1,0 @@
-package logic.usecases.tasksUseCases
-
-import logic.models.Task
-import logic.repositoies.TasksRepository
-
-class EditTaskUseCase(
-    private val tasksRepository: TasksRepository
-) {
-    operator fun invoke(task: Task): Boolean {
-        return tasksRepository.edit(task)
-    }
-}
