--- src/main/kotlin/logic/usecases/tasksUseCases/GetTaskByIdUseCase.kt
@@ -1,14 +1,0 @@
-package logic.usecases.tasksUseCases
-
-import logic.models.Task
-import logic.repositoies.TasksRepository
-import java.util.*
-
-class GetTaskByIdUseCase(
-    private val tasksRepository: TasksRepository
-) {
-
-    operator fun invoke(id: UUID): Task {
-        return tasksRepository.getById(id)
-    }
-}
